backend/requirements.txt
fastapi
uvicorn[standard]
moviepy
deepgram-sdk
supabase
python-dotenv
